requirements.txt
flask
flask-wtf 
selenium 
pillow
beautifulsoup4
requests
loguru